Finish character material gen and add Somnia mats
