+++ libs/gi/mats/src/executors/gen-mats/src/index.ts
@@ -1,0 +1,6 @@
+export type UpgradeCost = {
+  cost: number
+  items: ItemAmount[]
+}
+
+export type ItemAmount = { item: string; amount: number }
